requirements.txt
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.2.0
cycler==0.11.0
exceptiongroup==1.3.1
filelock==3.12.2
fonttools==4.47.0
fsspec==2023.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
Jinja2==3.1.2
joblib==1.5.3
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.4
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.8.1.78
packaging==23.2
pandas==2.3.3
pillow==10.2.0
pygame==2.6.1
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sympy==1.12
threadpoolctl==3.6.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typer-slim==0.21.0
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.6
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
